backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-localflavor==4.0
djangorestframework==3.16.0
pillow==11.2.1
python-stdnum==1.20
sqlparse==0.5.3
typing_extensions==4.13.1
